# api-server/openapi.yaml
openapi: 3.0.1
info:
  title: Visulate for Oracle
  version: 2.6.2
  description: |-
    Visulate for Oracle is a tool to catalog a company's Oracle databases. It
    provides read-only access to the data dictionary for one or more registered
    databases.

    It allows DDL (data definition language) statements to be downloaded for database
    objects and CSV file generation to export table contents.

    It can catalog Oracle database instances in preparation for a compliance audit.
    Identify Enterprise Edition feature and option usage. Find Enterprise Edition
    license instances that could be be downgraded to Standard Edition 2.

    The APIs + an associated UI provide an Oracle data dictionary browsing service to help
    data engineers understand the structure and dependencies in Oracle databases
    that they plan to migrate to the Cloud.

    ## MCP (Model Context Protocol) Support

    Visulate provides comprehensive MCP support through two service endpoints:

    **API Server (port 3000):** Database introspection and AI integration
    - `/mcp` - Protocol server and message handling
    - `/mcp/context/{db}` - Object context retrieval
    - `/mcp/search-objects/{db}` - Natural language object search
    - `/ai` - AI-powered analysis with context

    **Query Engine (port 5000):** Secure SQL execution
    - `/mcp-sql` - Secure credential management and SQL execution
    - Eliminates password exposure through secure token system
...

...

...
  contact:
...
  license:
...
servers:
...
paths:
  /endpoints/:
    summary: Database endpoints
    get:
      summary: List database endpoints
      parameters:
        - in: query
          name: status
          description: toggle result to retrieve valid/invalid endpoints
          required: false
          schema:
            type: string
            example: invalid
      description: >-
        Gets valid or invalid database endpoints as an object of name:value pairs.
        Response name is the endpoint name. The response value
        is the connect string for valid connections. Invalid connections are returned as
        and object with the connect string as the first item and error as the second.
      responses:
        200:
          description: Endpoint list
          content:
            application/json:
              schema:
                type: object
                properties:
                  endpoint:
                    type: string
                    example: oracle18XE
                  connectString:
                    type: string
                    example: db.visulate.net:1521/XEPDB1
                  status:
                    type: object
                    properties:
                      connectString:
                        type: string
                        example: db.visulate.net:1521/XEPDB1
                      error:
                        type: string
                        example: Get connection timed out after 5000 ms
...
  /find/{name}:
    get:
      summary: Find objects in multiple databases
      parameters:
        - in: path
          name: name
          description: database object name
          required: true
          schema:
            type: string
            example: dbms_output
      responses:
        200:
          description: Returns a list of objects found in each database
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FindResult"
...
  /ddl/{db}/{owner}/{type}:
    get:
      summary: Generate DDL
      description: Generate DDL for the specified owner and object type
      parameters :
        - in: path
          name: db
          description: database endpoint
          required: true
          schema:
            type: string
            example: oracle18XE
        - in: path
          name: owner
          description: database schema
          required: true
          schema:
            type: string
            example: WIKI
        - in: path
          name: type
          description: database object type
          required: true
          schema:
            type: string
            example: PACKAGE BODY
      responses:
        200:
          description: DDL output produced by DBMS_METADATA
          content:
            text/plain:
              schema:
                type: string
                example: |
                  CREATE TABLE "WIKI"."RNT_LOOKUP_TYPES" ...
        403:
          description: DDL generation not supported for Oracle internal objects
        404:
          description: Requested database was not found
  /ddl/{db}/{owner}/{type}/{name}/{status}:
    get:
      summary: Generate DDL
      description: Generate DDL for the specified owner, object type, name and status
      parameters :
        - in: path
          name: db
          description: database endpoint
          required: true
          schema:
            type: string
            example: oracle18XE
        - in: path
          name: owner
          description: database schema
          required: true
          schema:
            type: string
            example: WIKI
        - in: path
          name: type
          description: database object type
          required: true
          schema:
            type: string
            example: PACKAGE BODY
        - in: path
          name: name
          description: database object name or wildcard
          required: true
          schema:
            type: string
            example: RNT*
        - in: path
          name: status
          description: database object name or wildcard
          required: true
          schema:
            type: string
            example: VALID
      responses:
        200:
          description: DDL output produced by DBMS_METADATA
          content:
            text/plain:
              schema:
                type: string
                example: |
                  CREATE TABLE "WIKI"."RNT_LOOKUP_TYPES" ...
        403:
          description: DDL generation not supported for Oracle internal objects
        404:
          description: Requested database was not found
  /api/collection:
    post:
      summary: Retrieve details of related objects
      description: This endpoint retrieves details of related objects based on the provided input. Calls the database object report endpoint for each related object and returns the results.
      operationId: getRelatedObjectDetails
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/McpLinkExtractionPayload'
      responses:
        '200':
          description: An array containing the full QueryResult details for the base object and all related objects.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QueryResult"
        '400':
          description: Bad request
  /api/collection/{db}:
    post:
      summary: Database object collection
      description: Identify a collection of database objects matching one or more object paterns
      parameters:
        - in: path
          name: db
          description: database endpoint
          required: true
          schema:
            type: string
            example: oracle18XE
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                required: [owner, type, name, status]
                properties:
                  owner:
                    type: string
                    example: WIKI
                  type:
                    type: string
                    example: '*'
                  name:
                    type: string
                    example: RNT_MENUS*
                  status:
                    type: string
                    example: '*'

      responses:
        200:
          description: List of marching objects and their "uses" dependencies
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    OBJECT_ID:
                      type: integer
                      example: 22733
                    OBJECT_NAME:
                      type: string
                      example: DBMS_STANDARD
                    OBJECT_TYPE:
                      type: string
                      example: PACKAGE
                    OWNER:
                      type: string
                      example: SYS
                    REQUIRED_BY:
                      type: array
                      items:
                        type: object
                        properties:
                          OBJECT_ID:
                            type: integer
                            example: 22783
                          OBJECT_NAME:
                            type: string
                            example: RNT_MENUES_PKG
                          OBJECT_TYPE:
                            type: string
                            example: PACKAGE
                          OWNER:
                            type: string
                            example: WIKI
        404:
          description: Requested database was not found
...
  /ai:
    get:
      summary: Check if AI is enabled
      description: |
        Returns a JSON object indicating whether the AI feature is enabled.
      responses:
        200:
          description: AI feature status
          content:
            application/json:
              schema:
                type: object
                properties:
                  enabled:
                    type: boolean
                    example: true

    post:
      summary: (MCP) Generate text using Google AI
      description: >-
        Once you have gathered all necessary context with getContext, use this tool to generate the final, human-readable answer for the user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  required:
                    - message
                    - context
                  properties:
                    message:
                      type: string
                      example: "What is the structure of the users table?"
                    context:
                      $ref: '#/components/schemas/McpContextPayload'
                - type: object
                  required:
                    - jsonrpc
                    - method
                  properties:
                    jsonrpc:
                      type: string
                      enum:
                        - "2.0"
                      example: "2.0"
                    method:
                      type: string
                      example: "agent.getCapabilities"
                    id:
                      type: integer
                      example: 1
                    params:
                      type: object
      responses:
        200:
          description: Dynamic agent response. For standard chat, this is a text/plain stream. For A2A, this is an application/json payload.
          content:
            text/plain:
              schema:
                type: string
                example: "The users table contains details of users..."
            application/json:
              schema:
                type: object
                properties:
                  jsonrpc:
                    type: string
                    enum: ["2.0"]
                    example: "2.0"
                  id:
                    type: integer
                    example: 1
                  result:
                    type: object
                  error:
                    type: object
        503:
          description: Service unavailable
          content:
            text/plain:
              schema:
                type: string
                example: "Google AI key is not set"
  /mcp:
    get:
      summary: (MCP) Model Context Protocol Server
      description: >-
        Main endpoint for the Model Context Protocol (MCP) server implementation.
        This endpoint handles MCP client connections using Server-Sent Events (SSE) for real-time communication.
        Provides access to database introspection tools and Oracle query generation capabilities.
      responses:
        '200':
          description: SSE stream established for MCP communication
          content:
            text/event-stream:
              schema:
                type: string
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
    post:
      summary: (MCP) Model Context Protocol Request Handler
      description: >-
        Handles MCP protocol messages for database introspection and query generation.
        Supports tools for searching database objects, getting object context, and generating SQL queries.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: MCP protocol message
      responses:
        '200':
          description: MCP response
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Bad request
        '500':
          description: Server error
    delete:
      summary: (MCP) Cleanup MCP Session
      description: >-
        Cleanup endpoint for MCP sessions. Handles connection cleanup and resource management.
      responses:
        '200':
          description: Session cleaned up successfully
        '500':
          description: Server error
  /mcp/context/{db}:
    post:
      summary: (MCP) Get full context for a starting object
      description: >-
        After finding a specific object with searchObjects, use this tool to get its full context, including details of all related objects. After gathering the context, call generativeAI with the context and the original user message.
      parameters:
        - in: path
          name: db
          description: database endpoint
          required: true
          schema:
            type: string
            example: oracle18XE
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [owner, name, type]
              properties:
                owner:
                  type: string
                  example: RNTMGR2
                name:
                  type: string
                  example: RNT_ACCOUNTS_RECEIVABLE
                type:
                  type: string
                  example: TABLE
      responses:
        '200':
          description: The complete context payload, ready to be sent to the /ai endpoint.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/McpContextPayload'
        '404':
          description: The specified starting object was not found.
  /mcp/search-objects/{db}:
    post:
      summary: (MCP) Search for database objects
      description: >-
        Searches a specific database for objects matching a user's natural language query. Use this as the first step to find a starting point when the user's query is ambiguous. The tool will respond with an array of matching objects. Example response: [ { "OWNER": "RNTMGR2", "NAME": "RNT_TENANCY_AGREEMENT", "TYPE": "TABLE", "RELEVANCE_SCORE": 460 }, { "OWNER": "RNTMGR2", "NAME": "RNT_TENANCY_AGREEMENT_V", "TYPE": "VIEW", "RELEVANCE_SCORE": 230 } ] Select the object with the highest RELEVANCE_SCORE value. After using this tool, call getContext with the owner, name, and type of the highest scoring object.
      parameters:
        - in: path
          name: db
          description: database endpoint
          required: true
          schema:
            type: string
            example: oracle18XE
      requestBody:
        required: true
        description: The search terms and object types to query for.
        content:
          application/json:
            schema:
              type: object
              required: [search_terms, object_types]
              properties:
                search_terms:
                  type: array
                  items:
                    type: string
                  example: ["amount", "due", "rental", "agreement"]
                object_types:
                  type: array
                  items:
                    type: string
                  example: ["TABLE", "VIEW"]
      responses:
        '200':
          description: A ranked list of matching database objects.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    owner:
                      type: string
                      example: RNTMGR2
                    name:
                      type: string
                      example: RNT_ACCOUNTS_RECEIVABLE
                    type:
                      type: string
                      example: TABLE
                    relevance_score:
                      type: number
                      example: 95.5
        '400':
          description: Bad request, likely missing search_terms.
  /mcp/schema-summary/{db}:
    post:
      summary: (MCP) Get schema summary
      description: Returns a summary of object types and counts for a specific schema.
      parameters:
        - in: path
          name: db
          description: database endpoint
          required: true
          schema:
            type: string
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [owner]
            properties:
              owner:
                type: string
      responses:
        '200':
          description: Schema summary
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    name: { type: string }
                    type: { type: string }
                    rows: { type: integer }
                    comments: { type: string }
  /mcp/schema-relationships/{db}:
    post:
      summary: (MCP) Get schema relationships
      description: Returns a list of foreign key relationships for a schema. Metadata is cached for ERD generation.
      parameters:
        - in: path
          name: db
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [owner]
              properties:
                owner: { type: string }
      responses:
        '200':
          description: List of relationships or cache summary
  /mcp/schema-columns/{db}:
    post:
      summary: (MCP) Get schema columns
      description: Returns a list of columns for all tables in a schema. Metadata is cached for ERD generation.
      parameters:
        - in: path
          name: db
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [owner]
              properties:
                owner: { type: string }
      responses:
        '200':
          description: List of columns or cache summary
  /api/token:
    post:
      summary: Create credential token
      description: Proxies to query-engine to create a secure credential token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [database, username, password, session_id]
              properties:
                database: { type: string }
                username: { type: string }
                password: { type: string }
                session_id: { type: string }
      responses:
        '200':
          description: Token created
    delete:
      summary: Revoke session tokens
      description: Revokes all tokens for a session or a specific database.
      parameters:
        - in: query
          name: session_id
          required: true
          schema: { type: string }
        - in: query
          name: database
          required: false
          schema: { type: string }
      responses:
        '200':
          description: Tokens revoked
  /download/{sessionId}/{filename}:
    get:
      summary: Download file
      description: Serves a file from the download directory.
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: string }
        - in: path
          name: filename
          required: true
          schema: { type: string }
      responses:
...
components:
  securitySchemes:
    dbCustomAuth:
      type: apiKey
      in: header
      name: X-DB-Credentials
      description: Encoded database credentials. The value must be a base64-encoded string of 'username/password@db'.
...